refactor(mdx): overhaul MDX integration, components, and styling for improved UX
- restructure MDX components into a dedicated folder for maintainability
- enhance Callout component: better alignment, font, color schemes, and margin handling
- fix list rendering to show proper bullets/numbers in blog posts
- improve code block and inline code styling for readability and mobile responsiveness
- update syntax highlighting colors for better contrast in light mode
- add developer comments to global CSS for clarity
- ensure all custom MDX components are easily usable in .mdx files
- apply best practices for accessibility and visual consistency across light and dark themes
