+++ app/_components/mdx/code.js
@@ -1,0 +1,25 @@
+import { highlight } from "sugar-high";
+
+/**
+ * Code component for MDX
+ * @description This component is used to display code snippets with syntax highlighting.
+ * It uses the sugar-high library for syntax highlighting.
+ * @param {string} children - The code snippet to display.
+ * @param {object} props - Additional props to pass to the code element.
+ * @returns {JSX.Element} The Code component.
+ * @example
+ * <Code language="javascript">
+ *   const hello = "Hello, world!";
+ * </Code>
+ */
+export default function Code({ children, ...props }) {
+  const codeHTML = highlight(children);
+  return (
+    <code
+      className="block overflow-x-auto max-w-full"
+      // biome-ignore lint/security/noDangerouslySetInnerHtml: This is safe because the HTML is generated by sugar-high, which is a trusted library that sanitizes the code.
+      dangerouslySetInnerHTML={{ __html: codeHTML }}
+      {...props}
+    />
+  );
+}
